visualization/requirements.txt
Flask==1.0.3
Flask-RESTful==0.3.7
Werkzeug==0.15.6
openslide-python==1.1.1
Pillow==6.2.1
gunicorn==20.0.4
Flask-PyMongo==2.3.0
flask-cors==3.0.8
streaming_form_data==1.5.2
requests==2.22.0